Initialize git project
This project was created well before using source control.

All files are in the state that I left them years ago, with the exception of the solver in
which I commented out two lines yesterday that were stalling the solver
(I was trying at the time to animate the solver but couldn't get it to
work, and ended up breaking it in the process).

Coming back to this after so long, I see some warnings and other
potential problems that I may address.